+++ client/src/features/results/EventResults.tsx
@@ -59,7 +59,7 @@ const EventResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
 
   return (
     <StyledEntityResultsRow
-      borderTopLeftRadius={hasSimpleSearchQuery ? '0px' : undefined}
+      $borderTopLeftRadius={hasSimpleSearchQuery ? '0px' : undefined}
     >
       {(isSuccess || isError) && (
         <Col xs={12}>